feat: cjsToEsm.cjs ass solution
